adição da service de cliente, app e pagamento

+++ src/main/java/com/grupo5/monolitoclean/application/usecase/aplicativo/AtualizaCustoUC.java
@@ -2,6 +2,7 @@ package com.grupo5.monolitoclean.application.usecase.aplicativo;
 
 import com.grupo5.monolitoclean.application.dtos.AplicativoDTO;
 import com.grupo5.monolitoclean.domain.entities.AplicativoModel;
+import com.grupo5.monolitoclean.domain.services.AplicativoService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -14,7 +15,7 @@ public class AtualizaCustoUC {
         this.aplicativoService = aplicativoService;
     }
 
-    public AplicativoDTO run(long idAplicativo, double custoMensal){
+    public AplicativoDTO run(long idAplicativo, float custoMensal){
         AplicativoModel aplicativo =  aplicativoService.atualizaCusto(idAplicativo, custoMensal);
         return AplicativoDTO.fromModel(aplicativo);
     }
